refactor(issue-237): mover HorasDiscriminadas a PublicEvents
Payload de DiaCalculado. Los dos namespaces de ControlHoras que tenia
Contracts -- Eventos y ValueObjects -- colapsan en PublicEvents.ControlHoras.
