+++ src/Bitakora.ControlAsistencia.Contracts/Bitakora.ControlAsistencia.Contracts.csproj
@@ -15,6 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="ControlHoras\ValueObjects\" />
     <Folder Include="Empleados\ValueObjects\" />
   </ItemGroup>
