+++ src/Bitakora.ControlAsistencia.Contracts/ControlHoras/Eventos/DiaCalculado.cs
@@ -1,4 +1,4 @@
-using Bitakora.ControlAsistencia.Contracts.ControlHoras.ValueObjects;
+using Bitakora.ControlAsistencia.PublicEvents.ControlHoras;
 using Bitakora.ControlAsistencia.PublicEvents.Empleados;
 using Cosmos.EventDriven.Abstractions;
