+++ src/Bitakora.ControlAsistencia.PublicEvents/Bitakora.ControlAsistencia.PublicEvents.csproj
@@ -12,8 +12,5 @@
   <ItemGroup>
     <PackageReference Include="Cosmos.EventDriven.Abstractions" Version="2.3.1" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="ControlHoras\" />
-  </ItemGroup>
 
 </Project>
